feat: [2u-281] fix isQuery pending, utils, hooks
