+++ src/course-outline/enable-highlights-modal/EnableHighlightsModal.test.jsx
@@ -42,7 +42,7 @@ describe('<EnableHighlightsModal />', () => {
 
     const submitButton = getByRole('button', { name: messages.submitButton.defaultMessage });
     fireEvent.click(submitButton);
-    expect(onEnableHighlightsSubmitMock).toHaveBeenCalledTimes(1);
+    expect(onEnableHighlightsSubmitMock).toHaveBeenCalledTimes();
   });
 
   it('calls the close function when the "Cancel" button is clicked', () => {
