+++ src/course-outline/utils/getChecklistForStatusBar.js
